requirements.txt
motor==3.0.0
Pillow==9.1.1
pyaes==1.6.1
pymongo==4.1.1
Pyrogram==2.0.30
PySocks==1.7.1
PyYAML==6.0
TgCrypto==1.2.3
uvloop==0.16.0
